Fix tests; add module in package json in order to avoid renaming to mjs all files; revert prepare extension from mjs to js
