+++ bin/wasm-node/javascript/package.json
@@ -4,6 +4,7 @@
   "description": "Light client that connects to Polkadot and Substrate-based blockchains",
   "author": "Parity Technologies <admin@parity.io>",
   "license": "GPL-3.0-or-later WITH Classpath-exception-2.0",
+  "type": "module",
   "homepage": "https://github.com/paritytech/smoldot",
   "repository": {
     "type": "git",
@@ -24,10 +25,10 @@
   },
   "scripts": {
     "buildModules": "tsc -p tsconfig-mjs.json && tsc -p tsconfig-cjs.json && bash fix-package-type.sh",
-    "prepublishOnly": "node prepare.mjs --release && rimraf ./dist && npm run buildModules",
-    "build": "node prepare.mjs --release && rimraf ./dist && npm run buildModules",
-    "start": "node prepare.mjs --debug && rimraf ./dist && npm run buildModules && node demo/demo.mjs",
-    "test": "node prepare.mjs --debug && rimraf ./dist && npm run buildModules && ava --timeout=2m --concurrency 2 --no-worker-threads"
+    "prepublishOnly": "node prepare.js --release && rimraf ./dist && npm run buildModules",
+    "build": "node prepare.js --release && rimraf ./dist && npm run buildModules",
+    "start": "node prepare.js --debug && rimraf ./dist && npm run buildModules && node demo/demo.mjs",
+    "test": "node prepare.js --debug && rimraf ./dist && npm run buildModules && ava --timeout=2m --concurrency 2 --no-worker-threads"
   },
   "browser": {
     "./dist/mjs/compat/index.js": "./dist/mjs/compat/index-browser-overwrite.js",
