+++ bin/wasm-node/javascript/prepare.js
